ZERO-112 [Feat] : 관리자만 목록 삭제할 수 있도록 구현

+++ src/components/team/taskList/DeleteTaskListModal.tsx
@@ -4,6 +4,7 @@ import Button from '@components/@shared/Button';
 import { Modal } from '@components/@shared/Modal';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import Image from 'next/image';
+import { UserData } from '../member/ExileUserModal';
 
 interface DeleteTaskListModalProps {
   isOpen: boolean;
@@ -20,6 +21,14 @@ export default function DeleteTaskListModal({
 }: DeleteTaskListModalProps) {
   const queryClient = useQueryClient();
   const { id } = useTeamStore();
+  // 'user' 키로 캐싱된 유저 데이터 가져오기
+  const userData = queryClient.getQueryData<UserData>(['user']);
+
+  // 팀 관리자만 삭제 가능
+  const isAdmin =
+    userData &&
+    userData.memberships.find((m) => m.groupId === Number(id))?.role ===
+      'ADMIN';
 
   // 목록 삭제 Mutation
   const { mutate: deleteList } = useMutation({
@@ -37,6 +46,10 @@ export default function DeleteTaskListModal({
   });
 
   const handleDeleteClick = () => {
+    // 팀 관리자만 삭제 가능
+    if (!isAdmin) {
+      return;
+    }
     deleteList(taskListId);
   };
 
@@ -63,12 +76,22 @@ export default function DeleteTaskListModal({
             width={24}
             height={24}
           />
-          목록을 삭제하시겠어요?
+          {isAdmin ? (
+            <span>목록을 삭제하시겠어요?</span>
+          ) : (
+            <span>권한 없음</span>
+          )}
         </Modal.Header>
         <Modal.Content fontColor="secondary" fontSize="14" fontArray="center">
-          <p className="mt-[20px]">
-            [{taskName}] 내의 모든 할 일이 사라집니다. 정말로 삭제하시겠습니까?
-          </p>
+          {isAdmin && (
+            <p className="mt-[20px]">
+              [{taskName}] 내의 모든 할 일이 사라집니다. 정말로
+              삭제하시겠습니까?
+            </p>
+          )}
+          {!isAdmin && (
+            <p className="mt-[20px]">관리자만 삭제할 수 있습니다.</p>
+          )}
         </Modal.Content>
       </Modal.Wrapper>
       <Modal.Footer>
@@ -81,9 +104,11 @@ export default function DeleteTaskListModal({
           >
             취소
           </Button>
-          <Button size="full" bgColor="red" onClick={handleDeleteClick}>
-            삭제
-          </Button>
+          {isAdmin && (
+            <Button size="full" bgColor="red" onClick={handleDeleteClick}>
+              삭제
+            </Button>
+          )}
         </div>
       </Modal.Footer>
     </Modal>
